+++ pocketdesk/tests/android/system/StructStat.java
@@ -1,0 +1,9 @@
+package android.system;
+
+public final class StructStat {
+    public final int st_mode;
+
+    public StructStat(int mode) {
+        this.st_mode = mode;
+    }
+}